# PharmaVida-Usuarios/src/main/java/com/example/farmacy/security/dto/LoginResponseDto.java
package com.example.farmacy.security.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class LoginResponseDto {
    private String token;
}