# PharmaVida-Usuarios/src/main/java/com/example/farmacy/security/dto/RegisterResponseDto.java
package com.example.farmacy.security.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class RegisterResponseDto {
    private String token;
}